+++ R/ops_binary.R
@@ -12,41 +12,41 @@ NULL
 #' @rdname arith
 #' @export
 '+.uint8' <- function(e1, e2) {
-  uint8(do.call('+', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('+', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '-.uint8' <- function(e1, e2) {
-  uint8(do.call('-', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('-', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '*.uint8' <- function(e1, e2) {
-  uint8(do.call('*', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('*', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '/.uint8' <- function(e1, e2) {
-  uint8(do.call('/', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('/', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '%%.uint8' <- function(e1, e2) {
-  uint8(do.call('%%', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('%%', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '%/%.uint8' <- function(e1, e2) {
-  uint8(do.call('%/%', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('%/%', list(as.integer(e1), as.integer(e2))))
 }
 
 #' @rdname arith
 #' @export
 '^.uint8' <- function(e1, e2) {
-  uint8(do.call('^', list(as.integer(e1), as.integer(e2))))
+  as.uint8(do.call('^', list(as.integer(e1), as.integer(e2))))
 }