+++ R/uint8.R
@@ -1,7 +1,7 @@
 
 
 #-----------------------------------------------------------------------------
-#' Create a uint8 vector
+#' Creates objects of type 'uint8'
 #'
 #' The input is first converted to integer and this result (modulu 256) is converted
 #' to R's built-in \code{raw} datatype.
@@ -13,7 +13,7 @@
 #' @return Vector of \code{uint8} values.
 #' @export
 #-----------------------------------------------------------------------------
-uint8 <- function(x) {
+as.uint8 <- function(x) {
   i <- suppressWarnings(as.integer(x))
   if (any(is.na(i))) {
     stop("Non-integer values in:", paste(x, collapse=", "))
@@ -22,6 +22,27 @@ uint8 <- function(x) {
   attr(x, 'class') <- 'uint8'
   x
 }
+
+
+
+#-----------------------------------------------------------------------------
+#' Create a uint8 vector of the given size
+#'
+#' @param length A non-negative integer specifying the desired length.
+#'               Double values will be coerced to integer: supplying an argument
+#'               of length other than one is an error.
+#'
+#' @return A \code{uint8} vector of the specified length
+#' @export
+#-----------------------------------------------------------------------------
+uint8 <- function(length=0) {
+  x <- as.raw(integer(length))
+  attr(x, 'class') <- 'uint8'
+  x
+}
+
+
+
 
 
 #-----------------------------------------------------------------------------