+++ tests/testthat/test-sequence.R
@@ -1,8 +1,8 @@
 context("sequence")
 
 test_that("sequence works", {
-  b1 <- uint8(10)
-  b2 <- uint8(25)
+  b1 <- as.uint8(10)
+  b2 <- as.uint8(25)
   s1 <- b1:b2
   expect_s3_class(s1, 'uint8')
   expect_length(s1, 16)
@@ -11,7 +11,7 @@ test_that("sequence works", {
 
 
 test_that("order works", {
-  b <- uint8(c(10, 30, 20))
+  b <- as.uint8(c(10, 30, 20))
   o <- order(b)
   expect_type(o, 'integer')
   expect_identical(o, c(1L, 3L, 2L))