+++ vignettes/introduction.R
@@ -9,17 +9,23 @@ knitr::opts_chunk$set(
 )
 
 ## ------------------------------------------------------------------------
-b1 <- uint8(10)
-b2 <- uint8(20)
-b3 <- uint8(50)
+b02 <- as.uint8(  2)
+b0a <- as.uint8( 10)
+bff <- as.uint8(255)
 
-b1 + b2
 
-b3 * 5 + b2  #  270 %% 256 == 14
+# Addition is modulo 256
+bff + b0a   # (255 + 10) %% 256 => 9
 
-b3/b1
 
-b3/b2
+# Division result trucated to integer value
+b0a/b02   # 10/2 => 5
+bff/b0a   # 255/10 => 25.5 => 25
 
-b1 - b2  # -10  => 246
+
+# exponentiation is modulo 256
+b02 ^ b0a   # 2^10 => 1024.    1024 %% 256 => 0
+
+# initialisation is modulo 256
+as.uint8(seq(200, 300, 5))