+++ NEWS.md
@@ -1,5 +1,13 @@
 
 
+# uint8 0.1.1
+
+* 2018-05-01
+* Correct syntax to match other data creators i.e.
+    * `as.uint8(x)` converts given values into `uint8`
+    * `uint8(length)` creates a vector of the given length
+
+
 # uint8 0.1.0 
 
 * 2018-04-29